# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal, persiste en SQL Server 2022.",
        "title": "Galenos Pro Appointments API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pacientes": {
            "get": {
                "description": "Devuelve una página de pacientes ordenada, con datos de paginación.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Lista pacientes paginados",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tamaño de página (default 20, máximo 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/httpadapter.patientResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Persiste el paciente invocando el SP WebPacienteAgregar_E_H y devuelve el detalle del paciente creado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Registra un paciente nuevo",
                "parameters": [
                    {
                        "description": "Datos del paciente a registrar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.createPatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Paciente creado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpadapter.patientDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cuerpo inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al registrar el paciente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pacientes/buscar": {
            "get": {
                "description": "Invoca el procedimiento almacenado usp_go_listarpacientes. Los filtros vacíos se ignoran.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Busca pacientes aplicando filtros opcionales",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Número de documento del paciente",
                        "name": "documento",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Número de historia clínica",
                        "name": "hc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Apellido paterno",
                        "name": "paterno",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Apellido materno",
                        "name": "materno",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Primer nombre",
                        "name": "nombres",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Patient"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pacientes/por-documento": {
            "get": {
                "description": "Invoca el SP usp_go_ListarPacientePorNroDocyTipo con el número y el tipo de documento.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Busca un paciente por número y tipo de documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Número de documento del paciente",
                        "name": "nroDocumento",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id del tipo de documento de identidad",
                        "name": "idTipoDocIdentidad",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Patient"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Paciente no encontrado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pacientes/{idOrDoc}": {
            "get": {
                "description": "Si el path param es numérico invoca webPacientesListarIdPaciente (detalle); en caso contrario sp_listarPaciente.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Busca un paciente por id o por número de documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id o número de documento del paciente",
                        "name": "idOrDoc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PatientDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Identificador inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Paciente no encontrado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Patient": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "educationDegreeID": {
                    "type": "integer",
                    "format": "int64"
                },
                "firstName": {
                    "type": "string"
                },
                "historyNumber": {
                    "type": "string"
                },
                "homeAddress": {
                    "type": "string"
                },
                "homeCenterID": {
                    "type": "integer",
                    "format": "int64"
                },
                "homeDistrictID": {
                    "type": "integer",
                    "format": "int64"
                },
                "maritalStatusID": {
                    "type": "integer",
                    "format": "int64"
                },
                "maternalSurname": {
                    "type": "string"
                },
                "paternalSurname": {
                    "type": "string"
                },
                "patientID": {
                    "type": "integer",
                    "format": "int64"
                },
                "phone": {
                    "type": "string"
                },
                "secondName": {
                    "type": "string"
                },
                "sexTypeID": {
                    "type": "integer",
                    "format": "int64"
                },
                "thirdName": {
                    "type": "string"
                }
            }
        },
...
        "httpadapter.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpadapter.apiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/httpadapter.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpadapter.createPatientRequest": {
            "type": "object",
            "properties": {
                "apellidoMaterno": {
                    "type": "string"
                },
                "apellidoPaterno": {
                    "type": "string"
                },
                "direccionDomicilio": {
                    "type": "string"
                },
                "fechaNacimiento": {
                    "type": "string"
                },
                "idDistritoDomicilio": {
                    "type": "integer"
                },
                "idDistritoNacimiento": {
                    "type": "integer"
                },
                "idDocIdentidad": {
                    "type": "integer"
                },
                "idEstado": {
                    "type": "integer"
                },
                "idEstadoCivil": {
                    "type": "integer"
                },
                "idGradoInstruccion": {
                    "type": "integer"
                },
                "idProcedencia": {
                    "type": "integer"
                },
                "idTipoOcupacion": {
                    "type": "integer"
                },
                "idTipoSeguro": {
                    "type": "integer"
                },
                "idTipoSexo": {
                    "type": "integer"
                },
                "nroDocumento": {
                    "type": "string"
                },
                "nroHistoriaClinica": {
                    "type": "string"
                },
                "primerNombre": {
                    "type": "string"
                },
                "segundoNombre": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                }
            }
        },
...
        "httpadapter.patientResponse": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "educationDegreeId": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "historyNumber": {
                    "type": "string"
                },
                "homeAddress": {
                    "type": "string"
                },
                "homeCenterId": {
                    "type": "integer"
                },
                "homeDistrictId": {
                    "type": "integer"
                },
                "maritalStatusId": {
                    "type": "integer"
                },
                "maternalSurname": {
                    "type": "string"
                },
                "paternalSurname": {
                    "type": "string"
                },
                "patientId": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "secondName": {
                    "type": "string"
                },
                "sexTypeId": {
                    "type": "integer"
                },
                "thirdName": {
                    "type": "string"
                }
            }
        },
...
    }
}